requirements.txt
aiofiles
alphabet_detector
aiohttp
asyncio
apscheduler
beautifulsoup4
bs4
better_profanity
bleach
bing_image_downloader
CurrencyConverter
coffeehouse
cachetools
cloudscraper
countryinfo
dnspython
dateparser
emoji-country-flag
emoji
envparse
future
feedparser
faker
ffmpeg-python
fontTools
fuzzysearch
gtts
geopy
gpytranslate
googletrans==4.0.0rc1
google-trans-new
glitch_this
hachoir
heroku3
html2text
img2pdf
jikanpy
lxml
lyricsgenius==3.0.1
lyrics_extractor
markdown2
motor
multicolorcaptcha
nltk
nudepy
nekos.py
opencv-python-headless
odmantic
python-telegram-bot==13.11.0
psycopg2-binary
pyowm
pyaztro
pykeyboard
pynewtonmath
psutil
Pillow
pyrate-limiter
pretty_errors
pygithub
pyrogram==1.4.7
PyDictionary
pymongo
python-arq
pytz
pySmartDL
pytz
pendulum
requests
regex
redis
SQLAlchemy==1.4.31
spongemock
secureme
speedtest-cli
spamwatch
search_engine_parser
tswift
telethon
telegraph
TgCrypto
textblob 
tzlocal
ujson
wget
wikipedia
youtube_search_python==1.4.9
youtube_search
yt_dlp
youtube-dl
youtube_search
zalgo_text
git+https://github.com/Sorrybabe/downloader
git+https://github.com/python-telegram-bot/ptbcontrib.git
